Separates issuer from provider URL
The OpenID Provider URL and the issuer value are not necessarily the same
(see
http://openid.net/specs/openid-connect-core-1_0.html#IssuerIdentifier).
This change separates the provider URL from the issuer, and switches to
using the issuer returned from discovery, instead of the provider URL,
when checking the `iss` claim. The `issuer` value from the OpenID Provider metadata is a required value.